rework: player model and experience natives

+++ Un1t (second iteration)/Assets/Scripts/ExperienceComponent.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public class ExperienceComponent : MonoBehaviour
 {
-    [SerializeField] private float xp;
-    public float XP => xp;
+    [SerializeField] private int xp;
+    public int XP => xp;
 }
